+++ dotnet/examples/7_Performance.cs
@@ -121,7 +121,7 @@ namespace SEALNetExamples
             Random rnd = new Random();
             for (ulong i = 0; i < batchEncoder.SlotCount; i++)
             {
-                podValues[i] = (ulong)rnd.Next() % plainModulus.Value;
+                podValues[i] = plainModulus.Reduce(rnd.Next());
             }
 
             Console.Write("Running tests ");